Refactor Retryer
== DETAILS

Attempting to make the call method a little cleaner.

Also change the attempt number in Attempt from long to int, resolving
issue #27.

+++ src/main/java/com/github/rholder/retry/Attempt.java
@@ -74,7 +74,7 @@ public interface Attempt<T> {
      *
      * @return the attempt number
      */
-    long getAttemptNumber();
+    int getAttemptNumber();
 
     /**
      * The delay since the start of the first attempt, in milliseconds.
